fix: modularize CollectionAllocErr into errors.rs (addresses review feedback)
- Creates src/errors.rs with CollectionAllocErr enum and impls
- Removes inline definition from src/lib.rs
- Re-exports via pub use errors::CollectionAllocErr
- Addresses alejandro-vaz's review comment on #514
